% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_builder.R
\name{.gcv_select_lambda_for_dimension}
\alias{.gcv_select_lambda_for_dimension}
\title{Select optimal lambda using GCV for a dictionary dimension}
\usage{
.gcv_select_lambda_for_dimension(
  dictionary,
  dimension,
  embeddings_path,
  lambda_seq = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dictionary}{Data frame with 'word' column and dimension columns}

\item{dimension}{Name of the dimension to optimize}

\item{embeddings_path}{Path to embeddings file}

\item{lambda_seq}{Sequence of lambda values to test}

\item{verbose}{Print progress messages}
}
\value{
List with optimal_lambda and diagnostic information
}
\description{
Wrapper around .gcv_select_lambda that works with the vdic dictionary format.
Loads word vectors and applies GCV to select optimal regularization.
}
\keyword{internal}
