% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_vectionary.R
\name{.mean}
\alias{.mean}
\title{Compute mean metric (single document or vector)}
\usage{
.mean(text, word_projections, dimensions)
}
\arguments{
\item{text}{Character string or vector of strings}

\item{word_projections}{Data frame with word scores}

\item{dimensions}{Character vector of dimension names}
}
\value{
Named list with one element per dimension.
}
\description{
Compute mean metric (single document or vector)
}
\keyword{internal}
