% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_builder.R
\name{.project_all_embeddings_fast}
\alias{.project_all_embeddings_fast}
\title{Fast projection using data.table::fread (5-10x faster)}
\usage{
.project_all_embeddings_fast(axes, dimensions, embeddings_path, verbose = TRUE)
}
\description{
Fast projection using data.table::fread (5-10x faster)
}
\keyword{internal}
