% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_vectionary.R
\name{print.Vec-tionary}
\alias{print.Vec-tionary}
\title{Print method for Vec-tionary}
\usage{
\method{print}{`Vec-tionary`}(x, ...)
}
\arguments{
\item{x}{A vectionary object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for Vec-tionary
}
