% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_vectionary.R
\name{summary.Vec-tionary}
\alias{summary.Vec-tionary}
\title{Summary method for Vec-tionary}
\usage{
\method{summary}{`Vec-tionary`}(object, ...)
}
\arguments{
\item{object}{A vectionary object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summary method for Vec-tionary
}
